Changing the activation of inline editing http://code.google.com/a/eclipselabs.org/p/multiproperties/issues/detail?id=13
Currently the Alt+click activates the inline editing feature, which offers quite fast cell modification possibility.

A new toogle toolbar button should be introduced to let the user enable or disable the inline editing at all. When it is enabled a simple click is enough for activating the inline editing. The double click still activates the normal edit window.