http://code.google.com/a/eclipselabs.org/p/multiproperties/issues/detail?id=16 The command line usage requires too much class dependency
When the command line HandlerExecutor is try to load the handler, it tries to load Eclipse related classes which should be put on the classpath. There are so many of them in chain reaction, that the function is basically useless.

The basic cause of the problem is the WizardDialog and Shell classes.

The IHandler interface specifies the configure method with Shell parameter. Furthermore the method tries to create a WizardDialog.

Solution: 
The solution can be the create a new IHandlerConfigurer interface where the configure method is moved to. Thus the IHandler can use only Java specific classes only (no any Eclipse classes must be loaded).